Refactor resolver tests to enable end-to-end testing of execution.
