+++ src/TraceEvent/TraceEvent.Tests/AutomatedAnalysis/AnalyzerResolverTests.cs
@@ -8,13 +8,9 @@ using Xunit.Abstractions;
 using TestEventTests.Analyzers;
 using System.Linq;
 
-// Register the provider at the assembly level.
-[assembly: AnalyzerProvider(typeof(TestEventTests.Analyzers.TestAnalyzerProvider))]
-
 namespace TraceEventTests
 {
     [UseCulture("en-US")]
-
     public class AnalyzerResolverTests : TestBase
     {
         public AnalyzerResolverTests(ITestOutputHelper output)
@@ -44,7 +40,7 @@ namespace TraceEventTests
             SelfReportingAnalyzerResolver resolver = new SelfReportingAnalyzerResolver(AnalyzerSelector.AnalyzerOne);
             resolver.Resolve();
             Assert.Single(resolver.ResolvedAnalyzers);
-            Assert.Contains(TestAnalyzerProvider.One, resolver.ResolvedAnalyzers);
+            Assert.Contains(TestAnalyzerProvider.ResolverTests_AnalyzerOne, resolver.ResolvedAnalyzers);
         }
 
         [Fact]
@@ -53,7 +49,7 @@ namespace TraceEventTests
             SelfReportingAnalyzerResolver resolver = new SelfReportingAnalyzerResolver(AnalyzerSelector.AnalyzerTwo);
             resolver.Resolve();
             Assert.Single(resolver.ResolvedAnalyzers);
-            Assert.Contains(TestAnalyzerProvider.Two, resolver.ResolvedAnalyzers);
+            Assert.Contains(TestAnalyzerProvider.ResolverTests_AnalyzerTwo, resolver.ResolvedAnalyzers);
         }
 
         [Fact]
@@ -62,8 +58,8 @@ namespace TraceEventTests
             SelfReportingAnalyzerResolver resolver = new SelfReportingAnalyzerResolver(AnalyzerSelector.AnalyzerOne | AnalyzerSelector.AnalyzerTwo);
             resolver.Resolve();
             Assert.Equal(2, resolver.ResolvedAnalyzers.Count());
-            Assert.Contains(TestAnalyzerProvider.One, resolver.ResolvedAnalyzers);
-            Assert.Contains(TestAnalyzerProvider.Two, resolver.ResolvedAnalyzers);
+            Assert.Contains(TestAnalyzerProvider.ResolverTests_AnalyzerOne, resolver.ResolvedAnalyzers);
+            Assert.Contains(TestAnalyzerProvider.ResolverTests_AnalyzerTwo, resolver.ResolvedAnalyzers);
         }
     }
 
@@ -88,7 +84,7 @@ namespace TraceEventTests
         AnalyzerTwo
     }
 
-    public sealed class SelfReportingAnalyzerResolver : AnalyzerResolver
+    public sealed class SelfReportingAnalyzerResolver : TestAnalyzerResolver
     {
         private AnalyzerSelector _selector;
 
@@ -100,12 +96,12 @@ namespace TraceEventTests
         protected override void OnAnalyzerLoaded(AnalyzerLoadContext loadContext)
         {
             if (((_selector & AnalyzerSelector.AnalyzerOne) == AnalyzerSelector.AnalyzerOne) &&
-                (loadContext.Analyzer == TestAnalyzerProvider.One))
+                (loadContext.Analyzer == TestAnalyzerProvider.ResolverTests_AnalyzerOne))
             {
                 // Allow the analyzer to be run.
             }
             else if (((_selector & AnalyzerSelector.AnalyzerTwo) == AnalyzerSelector.AnalyzerTwo) &&
-                (loadContext.Analyzer == TestAnalyzerProvider.Two))
+                (loadContext.Analyzer == TestAnalyzerProvider.ResolverTests_AnalyzerTwo))
             {
                 // Allow the analyzer to be run.
             }
@@ -113,53 +109,6 @@ namespace TraceEventTests
             {
                 loadContext.ShouldRun = false;
             }
-        }
-
-        protected internal override void Resolve()
-        {
-            Assembly currentAssembly = Assembly.GetExecutingAssembly();
-            Assert.NotNull(currentAssembly);
-
-            ConsumeAssembly(currentAssembly);
-        }
-    }
-}
-
-namespace TestEventTests.Analyzers
-{
-    public sealed class TestAnalyzerProvider : IAnalyzerProvider
-    {
-        public static readonly AnalyzerOne One = new AnalyzerOne();
-        public static readonly AnalyzerTwo Two = new AnalyzerTwo();
-
-        internal static readonly Analyzer[] Analyzers = new Analyzer[]
-        {
-            One,
-            Two
-        };
-
-        IEnumerable<Analyzer> IAnalyzerProvider.GetAnalyzers()
-        {
-            foreach (Analyzer analyzer in Analyzers)
-            {
-                yield return analyzer;
-            }
-        }
-    }
-
-    public sealed class AnalyzerOne : Analyzer
-    {
-        protected override AnalyzerExecutionResult Execute(AnalyzerExecutionContext executionContext)
-        {
-            return AnalyzerExecutionResult.Success;
-        }
-    }
-
-    public sealed class AnalyzerTwo : Analyzer
-    {
-        protected override AnalyzerExecutionResult Execute(AnalyzerExecutionContext executionContext)
-        {
-            return AnalyzerExecutionResult.Success;
         }
     }
 }
